+++ include/Utilities/TimeHelper.hpp
@@ -3,12 +3,12 @@
 
 #include <chrono>
 
-inline long get_timestamp_sec() {
+inline int64_t get_timestamp_sec() {
   return std::chrono::duration_cast<std::chrono::seconds>(
       std::chrono::system_clock::now().time_since_epoch()).count();
 }
 
-inline unsigned long long get_timestamp_us() {
+inline int64_t get_timestamp_us() {
   return std::chrono::duration_cast<std::chrono::microseconds>(
       std::chrono::system_clock::now().time_since_epoch()).count();
 }
